requirements.txt
langchain==0.3.27
langchain-core~=0.3.75
sentence-transformers==5.1.0
langchain-qwq==0.2.1
pydantic~=2.11.7
python-dotenv==1.1.1
requests==2.32.5
